Update hex files and microbit source
This updates the hex files and the microbit source to the
versions run at Makespace for the demo event.

The privisioning is now handled by the bridge micro:bit

Beware: because the bridge was designed to be coded in blocks
it is naive, don't run the tonematrix program and the bridge
during provisioning! Start the matrix after provisioning